Remove redundant lesson repository
